+++ languagetool-core/src/test/java/org/languagetool/tools/ToolsTest.java
@@ -19,7 +19,12 @@
 package org.languagetool.tools;
 
 import org.junit.Test;
+import org.languagetool.JLanguageTool;
 import org.languagetool.TestTools;
+import org.languagetool.language.Demo;
+import org.languagetool.rules.CategoryId;
+import org.languagetool.rules.CategoryIds;
+import org.languagetool.rules.Rule;
 import org.languagetool.rules.RuleMatch;
 import org.languagetool.rules.patterns.PatternRule;
 import org.languagetool.rules.patterns.PatternToken;
@@ -27,9 +32,13 @@ import org.languagetool.rules.patterns.PatternToken;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import static org.hamcrest.core.Is.is;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 public class ToolsTest {
   
@@ -41,6 +50,42 @@ public class ToolsTest {
     match2.setSuggestedReplacement("I have");
     List<RuleMatch> matches = Arrays.asList(match1, match2);
     assertThat(Tools.correctTextFromMatches("I've have", matches), is("I've had"));
+  }
+
+  @Test
+  public void testSelectRules() {
+    Demo demo = new Demo();
+    expectDemoRuleId(Collections.emptySet(), Collections.emptySet(), Collections.emptySet(), Collections.emptySet(), false, demo);
+    expectNotDemoRuleId(Collections.emptySet(), Collections.emptySet(), Collections.singleton("DEMO_RULE"), Collections.emptySet(), false, demo);
+    expectNotDemoRuleId(Collections.singleton(CategoryIds.MISC), Collections.emptySet(), Collections.emptySet(), Collections.emptySet(), false, demo);
+    // disable category, but enable rule:
+    expectDemoRuleId(Collections.singleton(CategoryIds.MISC), Collections.emptySet(), Collections.emptySet(), Collections.singleton("DEMO_RULE"), false, demo);
+
+    // now with enablesOnly=true:
+    expectDemoRuleId(Collections.emptySet(), Collections.emptySet(), Collections.emptySet(), Collections.emptySet(), true, demo);
+    expectNotDemoRuleId(Collections.emptySet(), Collections.emptySet(), Collections.singleton("DEMO_RULE"), Collections.emptySet(), true, demo);
+    expectNotDemoRuleId(Collections.singleton(CategoryIds.MISC), Collections.emptySet(), Collections.emptySet(), Collections.emptySet(), true, demo);
+    expectDemoRuleId(Collections.emptySet(), Collections.singleton(CategoryIds.MISC), Collections.emptySet(), Collections.emptySet(), true, demo);
+    expectNotDemoRuleId(Collections.emptySet(), Collections.singleton(CategoryIds.CASING), Collections.emptySet(), Collections.emptySet(), true, demo);
+  }
+
+  private void expectDemoRuleId(Set<CategoryId> disabledCategories, Set<CategoryId> enabledCategories,
+                                Set<String> disabledRules, Set<String> enabledRules, boolean useEnabledOnly, Demo demo) {
+    JLanguageTool lt = new JLanguageTool(demo);
+    Tools.selectRules(lt, disabledCategories, enabledCategories, disabledRules, enabledRules, useEnabledOnly);
+    assertTrue(getRuleIds(lt).contains("DEMO_RULE"));
+  }
+
+  private void expectNotDemoRuleId(Set<CategoryId> disabledCategories, Set<CategoryId> enabledCategories,
+                                   Set<String> disabledRules, Set<String> enabledRules, boolean useEnabledOnly, Demo demo) {
+    JLanguageTool lt = new JLanguageTool(demo);
+    Tools.selectRules(lt, disabledCategories, enabledCategories, disabledRules, enabledRules, useEnabledOnly);
+    assertFalse(getRuleIds(lt).contains("DEMO_RULE"));
+  }
+
+  private List<String> getRuleIds(JLanguageTool lt) {
+    List<Rule> allActiveRules = lt.getAllActiveRules();
+    return allActiveRules.stream().map(Rule::getId).collect(Collectors.toList());
   }
 
   private static class FakeRule extends PatternRule {